requirements.txt
joblib
matplotlib
numpy
opencv-python
pillow
scikit-image
tensorflow-macos
tensorflow-metal
tifffile
tqdm